requirements.txt
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
rich>=13.0.0
python-dotenv
pydantic
mcp
PyYAML
